Split GPL code off into separate project not used by devices
The code that was left behind was code that I wrote myself, so I'm
relicensing that to Apache
