+++ Drivers/gps-generic-module/README.md
@@ -1,14 +1,10 @@
 # gps-generic-module
 
-Kernel driver implementing the `GPS_TYPE`/`GpsApi` interface with for generic UART-connected GPS/GNSS receivers:
-NMEA parsing for MTK, Airoha/AG33xx, ATGM336H/CASIC, Unicore UC6580 and u-blox 6/7/8/9/10 modules.
+This module contains only the generic interfaces to implement a driver.
 
-It is ported from [Meshtastic Firmware](https://github.com/MeshTastic/firmware), so it has a GPL v3.0 license.
+It serves as an interface for other GPS implementations, such as `gps-meshtastic-module`.
+The purpose is to provide an Apache-licensed interface while the implementation can be GPL-licensed.
 
 ## License
 
-This module is licensed under **GPL-3.0-or-later** (see `LICENSE-GPL-3.0.md`), separately from
-the rest of Tactility (Apache-2.0). The probing and initialization logic (`source/probe.cpp`,
-`source/init.cpp`, `source/ublox.cpp` and their private headers) is ported from
-[meshtastic/firmware](https://github.com/meshtastic/firmware) (GPL-3.0-or-later); see the
-`From: <url>` comments in those files for the exact origin of each ported function.
+This module is licensed under [Apache License v2.0](LICENSE-Apache-2.0.md).
