+++ Drivers/gps-meshtastic-module/CMakeLists.txt
@@ -6,6 +6,7 @@ file(GLOB_RECURSE SOURCE_FILES "source/*.c*")
 
 tactility_add_module(gps-generic-module
     SRCS ${SOURCE_FILES}
+    PRIV_INCLUDE_DIRS private/
     INCLUDE_DIRS include/
-    REQUIRES TactilityKernel gps-module minmea
+    REQUIRES TactilityKernel gps-module gps-generic-module minmea
 )
