+++ Drivers/gps-meshtastic-module/devicetree.yaml
@@ -1,0 +1,3 @@
+dependencies:
+  - TactilityKernel
+  - Modules/gps-module
