+++ Drivers/gps-meshtastic-module/include/gps_meshtastic/module.h
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+extern struct Module gps_meshtastic_module;
+
+#ifdef __cplusplus
+}
+#endif
