+++ Drivers/gps-meshtastic-module/private/gps_generic/private/gps_response.h
