+++ Drivers/gps-meshtastic-module/private/gps_generic/private/ublox_messages.h
